Adding relevant popup boxes

+++ client/src/main/java/client/scenes/EventOverviewCtrl.java
@@ -1,5 +1,7 @@
 package client.scenes;
 
+import client.dialog.ConfPopup;
+import client.dialog.Popup;
 import client.language.LanguageSwitch;
 import client.language.Translator;
 import client.utils.communicators.implementations.EventCommunicator;
@@ -299,13 +301,15 @@ public class EventOverviewCtrl implements Initializable, LanguageSwitch, SceneCo
         Optional<Participant> optionalParticipant = event.getParticipants().stream()
                 .filter(participant -> participant.getName().equals(participantDropDown.getValue()))
                 .findFirst();
-        // TODO: give a "confirmation" pop-up
-        if (optionalParticipant.isPresent()) {
-            participantCommunicator.deleteParticipant(optionalParticipant.get().getId());
-        }
-        // TODO: if no participant is selected
-        else {
-            System.out.println("Error");
+        boolean confirmed = ConfPopup.isConfirmed("Are you sure you want to delete this participant?");
+        if (confirmed) {
+            if (optionalParticipant.isPresent()) {
+                participantCommunicator.deleteParticipant(optionalParticipant.get().getId());
+            }else {
+                new Popup("No participant ID selected" , Popup.TYPE.ERROR).showAndWait();
+            }
+        } else {
+            return;
         }
         loadEvent(eventCommunicator.getEvent(event.getId()));
     }
@@ -349,10 +353,9 @@ public class EventOverviewCtrl implements Initializable, LanguageSwitch, SceneCo
                         if (updatedEvent != null && !updatedEvent.equals(event)) {
                             updateUI(updatedEvent);
                         }
-                        Thread.sleep(5000); // 5 seconds
                     }
-                } catch (InterruptedException e) {
-                    // Handle if the thread is interrupted
+                } catch (Exception e) {
+                    new Popup("Server polling error: " + e.getMessage(), Popup.TYPE.ERROR).showAndWait();
                 }
                 return null;
             }
@@ -377,7 +380,7 @@ public class EventOverviewCtrl implements Initializable, LanguageSwitch, SceneCo
      */
     private void updateUI(Event updatedEvent) {
         if (updatedEvent == null) {
-            System.err.println("The updated event is null, cannot update UI.");
+            new Popup("The updated event is null, cannot update UI.", Popup.TYPE.ERROR).showAndWait();
             return;
         }
 
@@ -401,7 +404,7 @@ public class EventOverviewCtrl implements Initializable, LanguageSwitch, SceneCo
             // Replace the local event object with the updated one
             this.event = updatedEvent;
 
-            System.out.println("UI has been updated with new event data.");
+            new Popup("Event update successful!", Popup.TYPE.INFO).show();
         });
     }
